requirements.txt
openai>=1.14.0
pandas>=2.2.0
psycopg2-binary>=2.9
pytest>=8.2.0
python-dotenv>=1.0
streamlit>=1.41.0
